fix mysqlcpp datetime

+++ mysqlcpp/QueryResult.h
@@ -79,7 +79,7 @@ public:
 
     uint64 getRowCount() const { return m_row_count; }
     uint32 getFieldCount() const { return m_field_count; }
-    ResultRow getRow(uint32 index) const;
+    ResultRow getRow(uint64 index) const;
 private:
     int nextRow();
     bool fetchRows();
